Convert type tests to use `tsd` And also move them closer to the package they test types of.
