--- integration-tests/types/index.d.ts
@@ -1,2 +1,0 @@
-// tslint:disable:no-useless-files
-// TypeScript Version: 3.3
