--- integration-tests/types/webclient-no-token.ts
@@ -1,11 +1,0 @@
-import { WebClient } from '@slack/web-api';
-
-const web = new WebClient();
-
-// $ExpectType Promise<WebAPICallResult>
-web.oauth.access({
-  client_id: 'CLIENT_ID',
-  client_secret: 'CLIENT_SECRET',
-  code: 'CODE',
-  redirect_uri: 'REDIRECT_URI',
-});
